ref(card): separating into its own file
